+++ cosmosid/api/auth.py
@@ -1,10 +1,8 @@
 import requests
 
 
-def get_profile(base_url, api_key):
-    response = requests.get(
-        f"{base_url}/api/auth/profile", json={}, headers={"X-Api-Key": api_key}
-    )
+def get_profile(base_url, headers):
+    response = requests.get(f"{base_url}/api/auth/profile", json={}, headers=headers)
 
     if response.status_code == 200:
         return response.json()