+++ cosmosid/api/import_workflow.py
@@ -2,8 +2,9 @@
 import json
 import logging
 
-import requests
-from requests import RequestException
+from cosmosid.helpers.exceptions import UploadException
+from requests import post
+from requests.exceptions import JSONDecodeError, RequestException, HTTPError
 
 LOGGER = logging.getLogger(__name__)
 
@@ -16,29 +17,41 @@ class ImportWorkflow(object):
 
     def import_workflow(self, workflow_ids, files, file_type, folder_id=None):
         upload_url = self.base_url + "/api/metagenid/v3/import"
-        payload = {}
-        payload["files"] = []
-        payload_dict = {
-            "metadata": {},
-            "parameters": {},
-            "source": "upload",
-            "folder_id": folder_id,
-            "workflow": workflow_ids,
-            "files": files["files"],
-            "sample_name": files["file_name"],
-            "sample_type": file_type,
+        payload = {
+            'files': [
+                {
+                    "metadata": {},
+                    "parameters": {},
+                    "source": "upload",
+                    "folder_id": folder_id,
+                    "workflow": workflow_ids,
+                    "files": files["files"],
+                    "sample_name": files["file_name"],
+                    "sample_type": file_type,
+                }
+            ]
         }
-        payload["files"].append(payload_dict)
         try:
-            response = requests.post(
+            response = post(
                 upload_url,
                 data=json.dumps(payload),
                 headers=self.header,
+
             )
+            errors = response.json().get('errors')
+            if errors:
+                raise UploadException(
+                    '\n'.join([
+                        f"Error with {error.get('sample', {}).get('sample_name')}: {error.get('description')}"
+                        for error in errors
+                    ])
+                )
             response.raise_for_status()
-        except requests.HTTPError as e:
-            print(f"{files} files can't be uploaded. Aborting.")
-            raise RuntimeError(e)
+        except HTTPError as err:
+            self.logger.error(f"{files} files can't be uploaded. Aborting.")
+            raise RuntimeError(err)
         except RequestException as err:
-            print("Upload request can't be send", err)
+            self.logger.error("Upload request can't be send", err)
+            raise RequestException(err)
+        except JSONDecodeError as err:
             raise RequestException(err)